update update installer & fix bug in installer addon

+++ SM64 ROM Manager.UpdateInstallerAddOn/AddOn.cs
@@ -5,20 +5,15 @@ namespace UpdateInstaller
 {
     public static class AddOn
     {
-        public static void Main(UpdateInstallerConfig config)
+        public static void Main(UpdateInstallerConfig config, bool isAfterCopyFiles)
         {
-            if (config.NewApplicationVersion is null && config.CurrentApplicationVersion is null)
+            if (!isAfterCopyFiles && config.NewApplicationVersion is null && config.CurrentApplicationVersion is null)
             {
                 foreach (string dir in Directory.GetDirectories(Path.Combine(config.HostApplicationPath, @"Data\Tweaks")))
                 {
-                    if (Path.GetFileName(dir) != "Reviewed")
-                    {
+                    if (Path.GetFileName(dir) == "Reviewed")
                         Directory.Delete(dir, true);
-                    }
                 }
-
-                foreach (string dir in Directory.GetFiles(Path.Combine(config.HostApplicationPath, @"Data\Tweaks")))
-                    File.Delete(dir);
             }
         }
     }
